making things simple
