+++ pytconf/convert.py
@@ -17,7 +17,6 @@ def convert_str_to_int(s: str) -> int:
     return int(s)
 
 
-# noinspection PyUnreachableCode
 def convert_str_to_int_default(i: int, s: str) -> int:
     if s.startswith("+"):
         return i + int(s[1:])
@@ -28,7 +27,6 @@ def convert_str_to_int_default(i: int, s: str) -> int:
     if s.startswith("/"):
         return i // int(s[1:])
     assert False, "you should not get here"
-    return 0
 
 
 def convert_int_to_str(i: int) -> str:
